requirements.txt
streamlit>=1.51.0
pandas
SQLAlchemy>=2.0.43
psycopg2-binary
Pillow
python-dotenv
